moved function to compute dice coefficient to utils.py
